./frontend/api for vercel root; vercel.json
I just saw Roshini's message.

"Backend" / API
---
This setting was already set in the vercel dashboard.
The "backend"/API used by vercel is in ./frontend/api.
The "cannonical" implementation is in ./modules/api.mjs.
Logically, the two APIs do the same thing.
The "cannonical" API was programmed for use with the node:http server
in ./modules/app.js and demuxed in ./modules/requests.mjs.
This commit is a logical equivalency with respect to the API.

SupabaseDB
---
Obviously, nothing has changed for this.
Any file that deals with our database has our supabase DB's API key
stored in plaintext as a constant near the top of the file (dig away!).
The two files that handle our API's POST requests at the /user endpoint
are the only two files with the API key: ./frontend/api/user.mjs and
./modules/supabase.mjs.

Other Comments
---
As I was skimming through the vercel docs for the `outputDirectory`
setting, I noticed that supabase was an integration in vercel -- oops!
I would like to rationalize our design decisions.
That is all.

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# On branch trunk
# Your branch is up to date with 'origin/trunk'.
#
# Changes to be committed:
#	new file:   vercel.json
#